Add track facade and remove service duplication
