+++ src/main/java/com/project/tracking_system/controller/HomeController.java
@@ -4,7 +4,7 @@ import com.project.tracking_system.dto.TrackInfoListDTO;
 import com.project.tracking_system.entity.Store;
 import com.project.tracking_system.entity.User;
 import com.project.tracking_system.service.store.StoreService;
-import com.project.tracking_system.service.track.TrackParcelService;
+import com.project.tracking_system.service.track.TrackFacade;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -30,7 +30,7 @@ import java.util.List;
 @RequestMapping("/")
 public class HomeController {
 
-    private final TrackParcelService trackParcelService;
+    private final TrackFacade trackFacade;
     private final StoreService storeService;
 
     /**
@@ -77,7 +77,7 @@ public class HomeController {
 
         try {
             // trackParcelService реализует логику с посылкой!
-            TrackInfoListDTO trackInfo = trackParcelService.processTrack(number, storeId, userId, canSave);
+            TrackInfoListDTO trackInfo = trackFacade.processTrack(number, storeId, userId, canSave);
 
             if (trackInfo == null || trackInfo.getList().isEmpty()) {
                 model.addAttribute("customError", "Нет данных для указанного номера посылки.");
